Split google auth ui logic into a seperate component
